refactor: separate prompt, cli, main module (#1)
ref #1
